Paired Sequence Retrieval re-write
One of the problems was that the paired sequence retrieval was not
working properly if the second pair was missing (this happens in an
error trimmed library where one pair may not be there because it got
removed)  aTRAM was acting strange if it did not find the second pair
and would print out no sequence but include a read name. We re-wrote
how it does the paired sequence retrieval to create two hash tables of
all the reads and only print out the ones where both reads are present.
All output files should have an even number of reads and an equal
number of pair 1 and pair 2 reads.