Two fixes with format_sra
on line 63 fixed an error in calculating file size when there are two
input files. This was affecting the number of shards the program would
produce if the same file was given as two input or one input file.

The second fix is allowing format_sra to be able to read input files
that have an _1 in the name rather than a /1 at the end of the name.
This has been checked for both 1 input file as well as for two
different input files.